moodlog/requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
passlib[bcrypt]
python-dotenv
openai
